% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm-functions.R
\name{osrmMatchBatch}
\alias{osrmMatchBatch}
\title{osrmMatchBatch function}
\usage{
osrmMatchBatch(x, y, timestamps, loc.id, trip.id)
}
\arguments{
\item{x}{Longitude coordinantes (degress).}

\item{y}{Latitude coordinantes (degress).}

\item{timestamps}{Array of timestamps (POSIXct) monotonically increasing.}

\item{loc.id}{GPS location identifiers (reference to the orginal points).}

\item{trip.id}{Trip identifiers.}
}
\value{
A list with two dataframes:
    (1) waypoints, location points that were snapped to the network.
    (2) matchings, network path including curves and corners.
}
\description{
It performs the map-matching in batch.
}
\examples{
# create a log directory in your working directory
ifelse(!dir.exists(file.path(getwd(), getOption("osrm.log"))),
        dir.create(file.path(getwd(), getOption("osrm.log"))),FALSE)
# run the batch process
out.mm <- osrmMatchBatch(ds3$longitude, ds3$latitude, ds3$timestamp,ds3$id, ds3$tripid)
}
\keyword{map-matching}

