% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm-functions.R
\name{osrmMatchBuildQuery}
\alias{osrmMatchBuildQuery}
\title{osrmMatchBuildQuery function}
\usage{
osrmMatchBuildQuery(x, y, timestamps, radiuses = NULL)
}
\arguments{
\item{x}{Longitude coordinantes (degress).}

\item{y}{Latitude coordinantes (degress).}

\item{timestamps}{Array of timestamps (POSIXct) monotonically increasing.}

\item{radiuses}{Standard deviation of GPS precision used for map matching. If applicable use GPS accuracy.}
}
\value{
a url request string
}
\description{
Build the request string for the match service in OSRM API v5.
}
\examples{
osrmMatchBuildQuery()
}
\keyword{map-matching}

