% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm-functions.R
\name{osrmMatchExtractMatchings}
\alias{osrmMatchExtractMatchings}
\title{osrmMatchExtractMatchings function}
\usage{
osrmMatchExtractMatchings(matchings)
}
\arguments{
\item{matchings}{dataframe with the road geometry to be parsed.}
}
\value{
dataframe with the road geometry
}
\description{
Parse the matching output into a dataframe.
}
\examples{
osrmMatchExtractMatchings()
}
\keyword{map-matching}

