% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm-functions.R
\name{segmentPartition}
\alias{segmentPartition}
\title{segmentPartition function}
\usage{
segmentPartition(dwell, timestamp, group)
}
\arguments{
\item{dwell}{Time in seconds for segmenting the timestamps.}

\item{group}{Array of the grouping variable (device) with same length as timestamps.}

\item{timestamps}{Array of timestamps (POSIXct) monotonically increasing.}
}
\value{
A dataframe with the ids (tripid)
}
\description{
It performs a segmentation using the dwell time as parameter.
}
\examples{
segmentPartition()
}
\keyword{dwell}
\keyword{segmentation}

